+++ frontend.js
@@ -1,13 +1,12 @@
-import {on_page_load} from './Lib/Dom.js';
-import {websocket} from './Lib/WebSocket.js'; 
+import {on_page_load} from './Lib/UI/Dom.js';
+import {websocket} from './Lib/UI/WebSocket.js'; 
 
-import { init_ui } from "./components.js";
+import { init_ui, websocket_class, set_ws } from "./components.js";
 
 
 let ws = new websocket().uri('localhost:9001')
-.on_message((event) => {
-	console.log('Message from server ', event.data);
-}).on_open(() => {
+.on_message(websocket_class.on_message())
+.on_open(() => {
 	console.log('WebSocket connection opened');
 }).on_close(() => {
 	console.log('WebSocket connection closed');
@@ -18,6 +17,7 @@ let ws = new websocket().uri('localhost:9001')
 on_page_load(() => {
 	console.log('Page loaded');
 
+	set_ws(ws);
 	init_ui();
 
 });