+++ Lib/UI/changelog.md
@@ -1,5 +1,15 @@
+# -01-05-2025 :
+- added `insert_child_at, insert_child_before, insert_child_after, selected, get_selected, first_child, get_first_child, last_child, get_last_child, checked, get_checked` methods to `element` class
+- added `label` function to `Dom.js` file
+- added `select, option` function to `Dom.js` file
+- added `Utils.js` file
+	- added `int, float, str, bool` functions
+
 # 30-04-2025 :
-- added `type, get_type` method to `element` class
+- added `type, get_type, attr, get_attr` method to `element` class
+- rename `get_child_by_index` to `get_child_at`, `remove_child_by_index` to `remove_child_at` methods in `element` class
+- added negative index support to `get_child_at` method in `element` class
+- added _ to all attribute members of `WebSocket` class
 
 # 22-04-2025 :
 - added `URL.js` file